refactor: usePreSeatGate를 entities로 이동해 FSD 역참조 해소
