+++ packages/client/src/entities/seat/model/usePreSeatGate.ts
@@ -1,14 +1,15 @@
-import { useIsPreSeatCrawlingPeriod } from '@/shared/config/preseat';
+import { useIsPreSeatCrawlingPeriod } from '@/entities/operationPeriod/model/usePreSeatPeriod';
 
-export const PRESEAT_START_TIME = '10:00';
-export const PRESEAT_CLOSE_DATE = 'operation-period 설정';
+interface IPreSeatGateOptions {
+  hasSeats?: boolean;
+}
 
 /**
  * preSeat의 가용성을 판단하는 훅입니다.
  * operation-period API에서 PRESEAT 기간을 조회하고
  * 기간 내 + 여석 데이터가 있으면 활성화합니다.
  */
-function usePreSeatGate(opts?: { hasSeats?: boolean }) {
+function usePreSeatGate(opts?: IPreSeatGateOptions) {
   const { hasSeats } = opts ?? {};
   const isCrawlingPeriod = useIsPreSeatCrawlingPeriod();
