+++ packages/client/src/pages/live/Live.tsx
@@ -7,7 +7,7 @@ import { Helmet } from 'react-helmet';
 import PinSearchBottomSheet from '@/widgets/live/pin/ui/PinSearchBottomSheet';
 import RealtimeTable from '@/widgets/live/board/ui/RealtimeTable';
 import useAlarmModalStore from '@/features/live/pin/model/useAlarmModalStore';
-import usePreSeatGate from '@/widgets/live/preSeat/model/usePreSeatGate';
+import usePreSeatGate from '@/entities/seat/model/usePreSeatGate';
 
 function Live() {
   const isSearchOpen = useAlarmModalStore(state => state.isSearchOpen);
