+++ packages/client/src/widgets/filtering/ui/WishFilter.tsx
@@ -12,7 +12,7 @@ import { FilterDomains, getCategories } from '@/features/filtering/lib/filterDom
 import FilteringButton from '@/features/filtering/ui/button/FilteringButton.tsx';
 import DepartmentFilter from '../../../features/filtering/ui/DepartmentFilter.tsx';
 import FilterDelete from '../../../features/filtering/ui/FilterDelete.tsx';
-import usePreSeatGate from '@/widgets/live/preSeat/model/usePreSeatGate.ts';
+import usePreSeatGate from '@/entities/seat/model/usePreSeatGate';
 import useWishesPreSeats from '@/entities/subjectAggregate/model/useWishesPreSeats.ts';
 import { IconButton, Flex, Chip } from '@allcll/allcll-ui';
 import { useWishesTableStore } from '@/features/wish/model/useWishTableColumnStore.ts';
